pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentos"
version = "0.1.0"
description = "The universal AI agent for your PC"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.30.0",
    "python-telegram-bot>=21.0",
    "rich>=13.7",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "cryptography>=42.0",
    # Phase 2: Screen control
    "mss>=9.0",
    "pyautogui>=0.9",
    "pynput>=1.7",
    "Pillow>=10.0",
    "open-clip-torch>=2.24",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]  # Line length handled by formatter; regex strings can't be split

[tool.ruff.lint.per-file-ignores]
"agentos/api/server.py" = ["B008"]  # FastAPI Depends() in function defaults is idiomatic

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
